Styling to give at least the appearance of some design
Since this project is close to Github, the more subtle links follow in
style, that is with inherited color and without underline on hover.
Futhermore, html5 has called and demanded semantic structure so did just
that. The author info is no longer a header and also appears outside the
direct text flow. No need to appear like savage coders without reason ;)